Move the 'mapping' models to their own mappings module, per the suggestion here: https://github.com/astropy/astropy/pull/3231#issuecomment-67657698  This also made a slight tweak to how the special method __call__ is handled in _ModelMeta, so that it is unnecessary for these models, that have undefined inputs/outputs until they are instantiated, to define dummy __call__ methods in order to work.
